Make sure cell is float

+++ atooms/system/cell.py
@@ -9,7 +9,7 @@ from atooms.core import ndim as _ndim
 class Cell(object):
 
     def __init__(self, side, origin=numpy.zeros(_ndim)):
-        self.side = numpy.asarray(side)
+        self.side = numpy.asarray(side, dtype=numpy.float64)
         self.origin = origin
 
     @property
